Added View Cart button.
