+++ config/sites/uipress.uiowa.edu/core.entity_view_display.node.book.default.yml
@@ -210,7 +210,7 @@ third_party_settings:
                 settings: {  }
                 third_party_settings: {  }
             additional: {  }
-            weight: 18
+            weight: 20
           -
             uuid: 21f2f637-bf22-4a1e-8f9c-1507246c4171
             region: first
@@ -225,7 +225,7 @@ third_party_settings:
                 settings: {  }
                 third_party_settings: {  }
             additional: {  }
-            weight: 21
+            weight: 23
           -
             uuid: 9e273933-3ad0-4fa8-aed3-43e0df81ad48
             region: first
@@ -240,7 +240,7 @@ third_party_settings:
                 settings: {  }
                 third_party_settings: {  }
             additional: {  }
-            weight: 20
+            weight: 22
           -
             uuid: f2701e59-27ca-4376-8f36-0a38d3025e79
             region: first
@@ -276,7 +276,7 @@ third_party_settings:
                 settings: {  }
                 third_party_settings: {  }
             additional: {  }
-            weight: 22
+            weight: 24
           -
             uuid: 7538c5e2-c8ff-404c-8153-e3e5276326ab
             region: first
@@ -375,7 +375,7 @@ third_party_settings:
                 settings: {  }
                 third_party_settings: {  }
             additional: {  }
-            weight: 19
+            weight: 21
           -
             uuid: b62eec0e-c50c-4cb4-9590-a555e4a3fe5c
             region: first
@@ -411,7 +411,7 @@ third_party_settings:
                 settings: {  }
                 third_party_settings: {  }
             additional: {  }
-            weight: 23
+            weight: 25
           -
             uuid: 333e4a7d-287f-4934-9966-1f5aeeb01899
             region: second
@@ -532,7 +532,7 @@ third_party_settings:
                 entity: layout_builder.entity
                 view_mode: view_mode
             additional: {  }
-            weight: 16
+            weight: 17
           -
             uuid: 44c4a146-ecfe-4c2d-bfbe-745fd0f692b4
             region: first
@@ -548,18 +548,18 @@ third_party_settings:
                   link: true
                 third_party_settings: {  }
             additional: {  }
-            weight: 17
+            weight: 18
           -
-            uuid: 764ce6c1-7358-455c-8cbf-940a1cc80d2d
+            uuid: 55445da0-716f-436d-b321-f23c580494ed
             region: first
             configuration:
-              id: addtocartbutton_block
+              id: cartbuttons_block
               label: null
               provider: uipress_core
               label_display: null
               context_mapping: {  }
             additional: {  }
-            weight: 15
+            weight: 19
         third_party_settings:
           layout_builder_lock:
             lock: {  }
